feat: Migrations & related checks can be disabled
The checks that verify the migration state are important to run the
application in a safe and sane way. However, in some emergency
circumstances it can be necessary to disable checks so that the
application may be started when the migration versions don't line up.

An example scenario might be an emergency release rollback.

Running in such a mode is inherently unsafe! It is the operator's
responsibility to decide if it is OK to run in this mode. The difference
between the target and existing migration state should be examined
carefully.

New environment variables:

- `MIGRATION_BYPASS_ENABLED`, boolean, not required, default: false

Signed-off-by: Ticean Bennett <ticean@reactioncommerce.com>